requirements.txt
altgraph==0.17.2
asgiref==3.8.1
Babel==2.15.0
backports.tarfile==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.7
decorator==5.1.1
Django==4.2.13
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.1
docopt==0.6.2
docutils==0.21.2
firebase-admin==6.5.0
future==0.18.2
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
install==1.3.5
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.4
keyring==25.2.1
lxml==5.2.2
lxml_html_clean==0.1.1
macholib==1.15.2
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
num2words==0.5.13
passlib==1.7.4
pillow==10.3.0
polib==1.2.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.0
requests==2.32.2
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
v==1
Werkzeug==3.0.3
zipp==3.18.2
reportlab==4.2.2
